Branch GlyphMetrics
